+++ intellij/create/src/NPC.java
@@ -12,7 +12,7 @@ public class NPC extends Entity {
         addVariable("health", "Integer", "50", false);
         addVariable("courage", "Integer", "10", false);
         addVariable("speed", "Integer", "4", false);
-        addVariable("dmgNoWeapon", "String", "1W3 - 1", false);
+        addVariable("dmgNoWeapon", "String", "1D3 - 1", false);
         addVariable("equippedWeapon", "String", "", false);
         addVariable("armor", "Integer", "0", false);
     }
@@ -48,34 +48,33 @@ public class NPC extends Entity {
     }
 
     public String generateSaveString() {
-        String str = "" + name + "\n" + description + "\n" + uid + "\n" + location + "\n" + inventory + "\n" + image;
+        StringBuilder str = new StringBuilder("" + name + "\n" + description + "\n" + uid + "\n" + location + "\n" + inventory + "\n" + image);
         for (int i = 0; i < eventName.size(); i++)
-            str = str + "\n++ev++" + eventName.get(i) + "---" + eventCode.get(i);
-        for (int i = 0; i < tags.size(); i++)
-            str = str + "\n++tag++" + tags.get(i);
+            str.append("\n++ev++").append(eventName.get(i)).append("---").append(eventCode.get(i));
+        for (String tag : tags) str.append("\n++tag++").append(tag);
         for (int i = 0; i < localVarName.size(); i++)
-            str = str + "\n++variable++" + localVarUids.get(i) + "---" + localVarName.get(i) + "---" + localVarType.get(i) + "---" + localVarValue.get(i) + "\n";
-        return str;
+            str.append("\n++variable++").append(localVarUids.get(i)).append("---").append(localVarName.get(i)).append("---").append(localVarType.get(i)).append("---").append(localVarValue.get(i)).append("\n");
+        return str.toString();
     }
 
     public String generateInformation() {
-        String str = "Name: " + name + "\nDescription: " + description + "\nLocation: " + location + "; " + Manager.getLocationName(location) +
-                "\nInventory: " + inventory + "; " + Manager.getInventoryName(inventory) + "\nImage: " + image + "; " + Manager.getImageName(image) + "\nEvents:";
+        StringBuilder str = new StringBuilder("Name: " + name + "\nDescription: " + description + "\nLocation: " + location + "; " + Manager.getLocationName(location) +
+                "\nInventory: " + inventory + "; " + Manager.getInventoryName(inventory) + "\nImage: " + image + "; " + Manager.getImageName(image) + "\nEvents:");
         for (int i = 0; i < eventName.size(); i++)
-            str = str + "\n   " + i + ": " + eventName.get(i);
-        str = str + "\nTags:\n";
+            str.append("\n   ").append(i).append(": ").append(eventName.get(i));
+        str.append("\nTags:\n");
         for (int i = 0; i < tags.size(); i++)
-            str = str + " " + i + ": " + tags.get(i) + ";";
-        str = str + "\nLocal variables:\n";
+            str.append(" ").append(i).append(": ").append(tags.get(i)).append(";");
+        str.append("\nLocal variables:\n");
         for (int i = 0; i < localVarName.size(); i++)
-            str = str + " " + localVarUids.get(i) + " - " + localVarName.get(i) + " - " + localVarType.get(i) + " - " + localVarValue.get(i) + "\n";
-        return str;
+            str.append(" ").append(localVarUids.get(i)).append(" - ").append(localVarName.get(i)).append(" - ").append(localVarType.get(i)).append(" - ").append(localVarValue.get(i)).append("\n");
+        return str.toString();
     }
 
     public void setLocation(String uid) {
         try {
             if (uid.equals("")) location = uid;
-        } catch (Exception e) {
+        } catch (Exception ignored) {
         }
         if (!StaticStuff.isValidUID(uid)) return;
         if (!Manager.locationExists(uid)) {